+++ upnp/src/genlib/net/http/httpreadwrite.c
@@ -630,7 +630,7 @@ int http_SendMessage(SOCKINFO *info, int *TimeOut, const char *fmt, ...)
 					rc = snprintf(Chunk_Header,
 						sizeof(Chunk_Header),
 						"%" PRIzx "\r\n",
-						(unsigned long)num_read);
+						num_read);
 					if (rc < 0 ||
 						(unsigned int)rc >=
 							sizeof(Chunk_Header)) {
@@ -1769,10 +1769,7 @@ int http_MakeMessage(membuffer *buf,
 		} else if (c == 'd') {
 			/* integer */
 			num = (size_t)va_arg(argp, int);
-			rc = snprintf(tempbuf,
-				sizeof(tempbuf),
-				"%" PRIzu,
-				(unsigned long)num);
+			rc = snprintf(tempbuf, sizeof(tempbuf), "%" PRIzu, num);
 			if (rc < 0 || (unsigned int)rc >= sizeof(tempbuf) ||
 				membuffer_append(buf, tempbuf, strlen(tempbuf)))
 				goto error_handler;
